Checkpoint: rename and move stuff around.
