+++ airbyte-workers/src/main/java/io/airbyte/workers/general/DbtTransformationRunner.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2021 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.workers.generalworkers;
+package io.airbyte.workers.general;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.google.common.base.Strings;
