+++ airbyte-workers/src/main/java/io/airbyte/workers/general/DefaultCheckConnectionWorker.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2021 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.workers.generalworkers;
+package io.airbyte.workers.general;
 
 import io.airbyte.commons.enums.Enums;
 import io.airbyte.commons.io.IOs;
@@ -16,8 +16,8 @@ import io.airbyte.protocol.models.AirbyteMessage;
 import io.airbyte.protocol.models.AirbyteMessage.Type;
 import io.airbyte.workers.*;
 import io.airbyte.workers.process.IntegrationLauncher;
-import io.airbyte.workers.protocols.airbyte.AirbyteStreamFactory;
-import io.airbyte.workers.protocols.airbyte.DefaultAirbyteStreamFactory;
+import io.airbyte.workers.internal.AirbyteStreamFactory;
+import io.airbyte.workers.internal.DefaultAirbyteStreamFactory;
 import java.io.InputStream;
 import java.nio.file.Path;
 import java.util.Optional;
