+++ airbyte-workers/src/main/java/io/airbyte/workers/general/DefaultNormalizationWorker.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2021 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.workers.generalworkers;
+package io.airbyte.workers.general;
 
 import io.airbyte.config.Configs.WorkerEnvironment;
 import io.airbyte.config.NormalizationInput;
