+++ airbyte-workers/src/main/java/io/airbyte/workers/internal/DefaultAirbyteDestination.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2021 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.workers.protocols.airbyte;
+package io.airbyte.workers.internal;
 
 import com.google.common.base.Charsets;
 import com.google.common.base.Preconditions;
