+++ airbyte-workers/src/main/java/io/airbyte/workers/internal/HeartbeatMonitor.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2021 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.workers.protocols.airbyte;
+package io.airbyte.workers.internal;
 
 import com.google.common.annotations.VisibleForTesting;
 import java.time.Duration;
