+++ airbyte-workers/src/main/java/io/airbyte/workers/internal/MessageTracker.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2021 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.workers.protocols.airbyte;
+package io.airbyte.workers.internal;
 
 import io.airbyte.config.FailureReason;
 import io.airbyte.config.State;
