+++ airbyte-workers/src/main/java/io/airbyte/workers/run/WorkerRun.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2021 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.workers.worker_run;
+package io.airbyte.workers.run;
 
 import io.airbyte.commons.functional.CheckedSupplier;
 import io.airbyte.config.JobOutput;
