+++ airbyte-workers/src/main/java/io/airbyte/workers/temporal/sync/DbtTransformationActivityImpl.java
@@ -16,8 +16,8 @@ import io.airbyte.config.persistence.split_secrets.SecretsHydrator;
 import io.airbyte.scheduler.models.IntegrationLauncherConfig;
 import io.airbyte.scheduler.models.JobRunConfig;
 import io.airbyte.scheduler.persistence.JobPersistence;
-import io.airbyte.workers.generalworkers.DbtTransformationRunner;
-import io.airbyte.workers.generalworkers.DbtTransformationWorker;
+import io.airbyte.workers.general.DbtTransformationRunner;
+import io.airbyte.workers.general.DbtTransformationWorker;
 import io.airbyte.workers.Worker;
 import io.airbyte.workers.WorkerApp;
 import io.airbyte.workers.WorkerConfigs;
