+++ airbyte-workers/src/test/java/io/airbyte/workers/DefaultNormalizationWorkerTest.java
@@ -19,7 +19,7 @@ import io.airbyte.workers.normalization.NormalizationRunner;
 import java.nio.file.Files;
 import java.nio.file.Path;
 
-import io.airbyte.workers.generalworkers.DefaultNormalizationWorker;
+import io.airbyte.workers.general.DefaultNormalizationWorker;
 import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
