+++ airbyte-workers/src/test/java/io/airbyte/workers/protocols/airbyte/AirbyteProtocolPredicateTest.java
@@ -8,6 +8,7 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import io.airbyte.commons.json.Jsons;
+import io.airbyte.workers.internal.AirbyteProtocolPredicate;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
