+++ airbyte-workers/src/test/java/io/airbyte/workers/protocols/airbyte/HeartbeatMonitorTest.java
@@ -13,6 +13,8 @@ import java.time.Duration;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
 import java.util.function.Supplier;
+
+import io.airbyte.workers.internal.HeartbeatMonitor;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
