+++ airbyte-workers/src/test/java/io/airbyte/workers/worker_run/TemporalWorkerRunFactoryTest.java
@@ -22,6 +22,8 @@ import io.airbyte.config.StandardSyncOutput;
 import io.airbyte.protocol.models.ConfiguredAirbyteCatalog;
 import io.airbyte.scheduler.models.Job;
 import io.airbyte.workers.WorkerConstants;
+import io.airbyte.workers.run.TemporalWorkerRunFactory;
+import io.airbyte.workers.run.WorkerRun;
 import io.airbyte.workers.temporal.TemporalClient;
 import io.airbyte.workers.temporal.TemporalResponse;
 import java.io.IOException;
