+++ airbyte-workers/src/test/java/io/airbyte/workers/worker_run/WorkerRunTest.java
@@ -14,6 +14,8 @@ import io.airbyte.workers.OutputAndStatus;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+
+import io.airbyte.workers.run.WorkerRun;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
